Split dbs_test.rb to whittle away at CI bottlenecks (#56336)
* Update dbs_test.rb estimated duration to reflect actual runtime

The CI test distributor uses these estimates to bin-pack tests across
parallel workers. dbs_test.rb was estimated at 82 seconds but actually
takes 20+ minutes due to process isolation and repeated app builds.

This causes the scheduler to underestimate the test's impact, leading
to CI timeouts when dbs_test.rb ends up in a bucket with other tests.

* Combine 7 ERB YAML parsing tests into one parameterized test

These tests all verify that db:create/db:drop work with various ERB
patterns in database.yml. Since each test runs in process isolation,
combining them into one test that iterates through fixtures saves 6
process forks and app builds (~2-3 minutes of CI time).

The test data is now organized in a frozen hash constant, making it
easier to maintain and extend.

* Extract PostgreSQL-specific tests to dbs_postgresql_test.rb

Move the three tests that require PostgreSQL to a separate file:
- db:create works when schema cache exists and database does not exist
- db:schema:cache:dump dumps virtual columns
- db:prepare creates test database if it does not exist

This isolation allows:
1. CI to run SQLite tests without waiting for PostgreSQL tests
2. Better parallelization across CI workers
3. Clearer separation of database-specific test concerns

* Split dbs_test.rb into focused test files for better CI parallelization

The original dbs_test.rb contained 40 tests covering multiple areas of
database rake tasks. By splitting into focused files, CI can distribute
tests more evenly across parallel workers:

- dbs_test.rb: Core db:create/db:drop tests (9 tests)
- dbs_schema_test.rb: Schema dump/load/cache and migration tests (19 tests)
- dbs_setup_test.rb: Setup, prepare, seed, and truncate tests (12 tests)
- dbs_postgresql_test.rb: PostgreSQL-specific tests (3 tests, from previous commit)

This allows the CI scheduler to parallelize what was previously a single
20+ minute bottleneck across multiple workers.

* Add estimated durations for new dbs test files

Update the CI scheduler estimates to reflect the split test files:
- dbs_test.rb: 300s (9 tests, down from 1200s)
- dbs_schema_test.rb: 500s (19 tests)
- dbs_setup_test.rb: 350s (12 tests)
- dbs_postgresql_test.rb: 150s (3 tests)

These estimates enable the CI bin-packing algorithm to distribute
the tests across parallel workers more effectively.
