+++ railties/Rakefile
@@ -21,7 +21,10 @@ namespace :test do
       "test/application/test_test.rb" => 25,
       "test/application/configuration_test.rb" => 106,
       "test/generators/app_generator_test.rb" => 150,
-      "test/application/rake/dbs_test.rb" => 82,
+      "test/application/rake/dbs_test.rb" => 300,
+      "test/application/rake/dbs_schema_test.rb" => 500,
+      "test/application/rake/dbs_setup_test.rb" => 350,
+      "test/application/rake/dbs_postgresql_test.rb" => 150,
       "test/application/rake_test.rb" => 33,
       "test/generators/plugin_generator_test.rb" => 37,
       "test/railties/engine_test.rb" => 39,
